tests/submission/Infraware/Forms/contents/Forms/keygen_keytype.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>keygen_keytype</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form onsubmit="return false" name='input_form'>
	 <p><keygen name="key" id='keygen_id'></keygen></p>
	</form>

	<script>
		
		var keygen = document.getElementById("keygen_id");
		var keytype = keygen.keytype;

		test(function() {
			assert_equals(typeof(keytype), "string", "keytype attribute is not correct.");
		});

	</script>

 </body>
</html>
